requirements.txt
fastapi
uvicorn
nest_asyncio
transformers
torch
opencv-python-headless
Pillow
sqlalchemy
faiss-cpu
streamlit
tqdm
ultralytics

